backend/requirements.txt
# run 'pip install -r requirements.txt' to install dependencies
passlib
mysql-connector
argon2_cffi
PyJWT
flask
mock
jsonschema
flask-cors
pytest-cov
boto3
pylint
